pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ast-edit"
version = "0.1.0"
description = "Infrastructure for normalized code-edit datasets and ast-grep agents."
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2,<3",
    "huggingface-hub>=0.24",
    "requests>=2.31",
    "pyyaml>=6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
